Moved grabData function into a method so that it can be called multiple times. It can be found under the main method now.
Also, added a Timer function which can update Fitbit data using grabData
method. Currently, the Timer will grab data 5 times at 5 second
intervals. This is for testing purposes only so that we don't need to
wait hours for the program to complete. The test loop will create a txt
file to keep track of how many times data was updated.